.profile: Source .bashrc if running zsh
I don't know if the entirety of the .bashrc file will work in zsh, but
it should work in most login shell scenarios (that is, lightly
redirecting to fish when it exists). This hack is easier than changing
the login shell from zsh to (the ancient built-in version of) bash on
every Mac computer, probably.

Still not sure if macOS likes using something from Homebrew as a
full-blown login shell, so I continue to avoid such an approach.