fish_color_host: Attempt to support per-machine overrides
Try to respect a custom universal variable already set under that
variable name, if it exists. However, fish shell sets a universal
value for that variable already, so it's a bit gross.

Try to un-hardcode the function location, by using `status -f` to get
the directory of the config.fish file.