Shell tweaks to not assume bash so much
Use dash for a few things in bin/, because hopefully it'll be faster for
those.

Tell vim that we are always using a POSIX shell.

Update .profile to use $() instead of ``, as the former apparently
doesn't work in some ancient shells like OG Bourne shell, and that's
what vim is assuming as the lowest common denominator by default.