multi_sed: sed -i is not portable
Turns out the BSD and GNU implementations of `sed -i` are
irreconcilable, because it's not part of the POSIX standard. Explicitly
dump into a temporary file for now.

This script should really use a more efficient single-pass approach over
the files, at least for the non-Vim cases. (And the Vim case is bad
anyways, because it doesn't gracefully support early termination.)