fish_conf: Move all large functions out of config.fish
I doubt that the few extra disk reads will tank performance, especially
since a lot of the builtins are also autoload .fish functions anyways.

In terms of disk wear, hopefully the SSD controller / cache is smart
enough to not care about the file count when `git pull`-ing. And like,
git dumps a bunch more files in .git anyways, so ¯\_(ツ)_/¯